+++ src/components/Accounts/ViewFeesFines/ViewFeesFines.js
@@ -5,7 +5,6 @@ import {
   Button,
   Dropdown,
   DropdownMenu,
-  MenuItem,
   MultiColumnList,
 } from '@folio/stripes/components';
 
@@ -24,6 +23,7 @@ import {
   isRefundAllowed,
   isCancelAllowed,
 } from '../accountFunctions';
+
 
 class ViewFeesFines extends React.Component {
   static propTypes = {
@@ -262,14 +262,11 @@ class ViewFeesFines extends React.Component {
     (accounts.findIndex(a => a.id === f.id) !== -1);
   };
 
-  handleOptionsChange(itemMeta, e) {
-    e.preventDefault();
-    e.stopPropagation();
-
-    const { a, action } = itemMeta;
+  handleOptionsChange(itemMeta) {
+    const { account, action } = itemMeta;
 
     if (action && this[action]) {
-      this[action](a);
+      this[action](account);
     }
   }
   // ellipsis actions
@@ -316,59 +313,88 @@ class ViewFeesFines extends React.Component {
     nav.onClickViewLoanActionsHistory(e, { id: a.loanId }, history, params);
   }
 
+  /**
+   * renderToggle
+   * trigger for the ellipses menu
+   */
+  renderToggle = ({ triggerRef, onToggle, ariaProps, keyHandler }) => (
+    <Button
+      data-test-ellipsis-button
+      ref={triggerRef}
+      onClick={onToggle}
+      onKeyDown={keyHandler}
+      buttonStyle="hover dropdownActive"
+      {...ariaProps}
+    >
+      <strong>•••</strong>
+    </Button>
+  );
+
+  /**
+   * MenuButton
+   * inner-class for an element on the ellipses-menu
+   */
+  MenuButton = ({ disabled, account, action, children }) => {
+    const onClick = (e) => {
+      e.preventDefault();
+      e.stopPropagation();
+      this.handleOptionsChange({ account, action });
+    };
+
+    return (
+      <Button disabled={disabled} buttonStyle="dropdownItem" onClick={onClick}>
+        {children}
+      </Button>
+    );
+  };
+
+  /**
+   * renderActions
+   * return the ellipses menu, a <Dropdown>
+   * @param a object: an account
+   */
   renderActions(a) {
     const { feeFineActions = [] } = this.props;
-    const disabled = (a.status.name === 'Closed');
-    const elipsis = {
-      pay: disabled,
-      waive: disabled,
-      transfer: disabled,
-      error: disabled || !isCancelAllowed(a),
+
+    // disable ellipses menu actions based on account-status
+    const isClosed = (a.status.name === 'Closed');
+    const isDisabled = {
+      pay: isClosed,
+      waive: isClosed,
+      transfer: isClosed,
+      error: isClosed || !isCancelAllowed(a),
       loan: (a.loanId === '0' || !a.loanId),
       refund: !isRefundAllowed(a, feeFineActions),
     };
 
+    // disable ellipses menu actions based on permissions
     const buttonDisabled = !this.props.stripes.hasPerm('ui-users.feesfines.actions.all');
 
     return (
       <Dropdown
-        onSelectItem={this.handleOptionsChange}
+        renderTrigger={this.renderToggle}
+        usePortal
       >
-        <Button data-test-ellipsis-button data-role="toggle" buttonStyle="hover dropdownActive">
-          <strong>•••</strong>
-        </Button>
-        <DropdownMenu id="ellipsis-drop-down" data-role="menu">
-          <MenuItem itemMeta={{ a, action: 'pay' }}>
-            <Button disabled={!((elipsis.pay === false) && (buttonDisabled === false))} buttonStyle="dropdownItem">
-              <FormattedMessage id="ui-users.accounts.history.button.pay" />
-            </Button>
-          </MenuItem>
-          <MenuItem itemMeta={{ a, action: 'waive' }}>
-            <Button disabled={!((elipsis.waive === false) && (buttonDisabled === false))} buttonStyle="dropdownItem">
-              <FormattedMessage id="ui-users.accounts.history.button.waive" />
-            </Button>
-          </MenuItem>
-          <MenuItem itemMeta={{ a, action: 'refund' }}>
-            <Button disabled={!((elipsis.refund === false) && (buttonDisabled === false))} buttonStyle="dropdownItem">
-              <FormattedMessage id="ui-users.accounts.history.button.refund" />
-            </Button>
-          </MenuItem>
-          <MenuItem itemMeta={{ a, action: 'transfer' }}>
-            <Button disabled={!((elipsis.transfer === false) && (buttonDisabled === false))} buttonStyle="dropdownItem">
-              <FormattedMessage id="ui-users.accounts.history.button.transfer" />
-            </Button>
-          </MenuItem>
-          <MenuItem itemMeta={{ a, action: 'cancel' }}>
-            <Button disabled={!((elipsis.error === false) && (buttonDisabled === false))} buttonStyle="dropdownItem">
-              <FormattedMessage id="ui-users.accounts.button.error" />
-            </Button>
-          </MenuItem>
+        <DropdownMenu id="ellipsis-drop-down">
+          <this.MenuButton disabled={isDisabled.pay || buttonDisabled} account={a} action="pay">
+            <FormattedMessage id="ui-users.accounts.history.button.pay" />
+          </this.MenuButton>
+          <this.MenuButton disabled={isDisabled.waive || buttonDisabled} account={a} action="waive">
+            <FormattedMessage id="ui-users.accounts.history.button.waive" />
+          </this.MenuButton>
+          <this.MenuButton disabled={isDisabled.refund || buttonDisabled} account={a} action="refund">
+            <FormattedMessage id="ui-users.accounts.history.button.refund" />
+          </this.MenuButton>
+          <this.MenuButton disabled={isDisabled.transfer || buttonDisabled} account={a} action="transfer">
+            <FormattedMessage id="ui-users.accounts.history.button.transfer" />
+          </this.MenuButton>
+          <this.MenuButton disabled={isDisabled.error || buttonDisabled} account={a} action="cancel">
+            <FormattedMessage id="ui-users.accounts.button.error" />
+          </this.MenuButton>
           <hr />
-          <MenuItem itemMeta={{ a, action: 'loanDetails' }}>
-            <Button disabled={elipsis.loan} buttonStyle="dropdownItem">
-              <FormattedMessage id="ui-users.accounts.history.button.loanDetails" />
-            </Button>
-          </MenuItem>
+          <this.MenuButton disabled={isDisabled.loan || buttonDisabled} account={a} action="loanDetails">
+            <FormattedMessage id="ui-users.accounts.history.button.loanDetails" />
+          </this.MenuButton>
         </DropdownMenu>
       </Dropdown>
     );